chore(pong): rename rendering files
