+++ frontend/src/js/game/rendering.js
